feat/todo-app#32: refactor structure
