+++ frontend/src/app/login/_config/login.schema.ts
@@ -1,31 +1,9 @@
+import { emailSchema } from "@/types/validation"
 import { z } from "zod"
-
-export const emailSchema = z
-  .string()
-  .trim()
-  .min(1, "Email is required")
-  .email("Invalid email address")
 
 export const loginPasswordSchema = z
   .string()
   .min(1, "Password is required")
-
-export const registerPasswordSchema = z
-  .string()
-  .min(8, "Password must be at least 8 characters")
-  .regex(
-    /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)/,
-    "Password must include uppercase, lowercase, and a number"
-  )
-
-export const fullNameSchema = z
-  .string()
-  .trim()
-  .min(2, "Name must be at least 2 characters")
-  .max(100, "Name must be less than 100 characters")
-
-
-// LOGIN
 
 export const loginSchema = z.object({
   email: emailSchema.transform((v) => v.toLowerCase()),
@@ -46,14 +24,12 @@ export const loginResponseSchema = z.object({
 export type TLoginResponseDto = z.infer<typeof loginResponseSchema>
 
 // AUTH SUCCESS RESPONSE (COMBINED)
-
 export type TAuthSuccessResponse = {
   accessToken: string
   refreshToken: string
 }
 
 // AUTH ERROR TYPE
-
 export type TAuthApiError = {
   message: string
   code?: string
