+++ frontend/src/types/validation.ts
@@ -1,0 +1,7 @@
+import z from "zod";
+
+export const emailSchema = z
+  .string()
+  .trim()
+  .min(1, "Email is required")
+  .email("Invalid email address")
